requirements.txt
numpy
pandas
opencv_contrib_python
matplotlib
Pillow